Split project into vaadin app and core submodules
The core submodule represents the vaadin-independent backend and the app
submodule contains the vaadin web app. This was done to allow future
plugins to only depend on the core module, and also, its just a lot
cleaner (similarly, if one day one decides to create a different UI than
the Vaadin app, like a native mobile app etc., one can just reuse the
backend module and implement a new UI/rest interface/... for it).