Clean up session management logic
Now, the MapSessionManager is responsible for keeping the invariant that
a player is always in exactly one session, and also keeps track of which
session each player is in. A player will now stay in a session until
they explicitly leave, or until the application is restartet (at which
point all sessions are lost). The user class now uses a
GameSessionManager (interface of MSM) to implement the user-specific
join/leave/getSession operations defined in PlayerSessionManager. The
GSM must be set from the outside, which currently happens in
AuthenticatedUser where Spring Autowiring is used for automatic
singleton-style creation and injection of a GameSessionManager (using a
spring component wrapper for MapSessionManager).

TODO: Check need for thread safety, as non of this is currently thread
safe (which it probably should be). Also, make a pretty UI for the
feature.